Use RunE for cobra commands
This gives us more control of the output.

In cases where we supply stdOut and stdErr (test cases), if
we abort the execution upon an error, the caller does not
have the chance to collect the output and print to user.

By using `RunE` instead of `Run` we don't need to abort
the execution and therefore break the workflow. By returning
an error the caller can properly handle the error.